Renamed some things
- Parser => BaseParser
- AsyncParser => Parser
