+++ core/src/main/scala/tectonic/json/Parser.scala
@@ -40,613 +40,348 @@ package json
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
+ * kDEALINGS IN THE SOFTWARE.
  */
 
 import tectonic.util.BList
 
-import scala.{inline, sys, Array, Boolean, Char, Int, Long, Nothing, Predef, Unit}, Predef._
-import scala.annotation.{switch, tailrec}
+import scala.{inline, Array, Boolean, Byte, Char, Either, Int, Left, Long, Right, Unit}
+import scala.annotation.switch
+import scala.math.max
+import scala.util.control
 
-import java.lang.{CharSequence, Exception, IndexOutOfBoundsException, String, SuppressWarnings}
-import java.nio.charset.Charset
+import java.lang.{CharSequence, Exception, String, SuppressWarnings, System}
+import java.nio.ByteBuffer
 
-final case class ParseException(msg: String, index: Int, line: Int, col: Int) extends Exception(msg)
+object Parser {
 
-final case class IncompleteParseException(msg: String) extends Exception(msg)
+  sealed abstract class Mode(val start: Int, val value: Int)
+  case object UnwrapArray extends Mode(-5, 1)
+  case object ValueStream extends Mode(-1, 0)
+  case object SingleValue extends Mode(-1, -1)
+
+  @SuppressWarnings(
+    Array(
+      "org.wartremover.warts.DefaultArguments",
+      "org.wartremover.warts.Null"))
+  def apply[A](plate: Plate[A], mode: Mode = SingleValue): Parser[A] =
+    new Parser(plate, state = mode.start, curr = 0,
+      ring = 0L, roffset = -1, fallback = null,
+      data = new Array[Byte](131072), len = 0, allocated = 131072,
+      offset = 0, done = false, streamMode = mode.value)
+}
 
 /**
- * Parser implements a state machine for correctly parsing JSON data.
+ * Parser is able to parse chunks of data (encoded as
+ * Option[ByteBuffer] instances) and parse asynchronously. You can
+ * use the factory methods in the companion object to instantiate an
+ * async parser.
  *
- * The trait relies on a small number of methods which are left
- * abstract, and which generalize parsing based on whether the input
- * is in Bytes or Chars, coming from Strings, files, or other input.
- * All methods provided here are protected, so different parsers can
- * choose which functionality to expose.
+ * The async parser's fields are described below:
  *
- * Parser is parameterized on J, which is the type of the JSON AST it
- * will return. Jawn can produce any AST for which a Facade[J] is
- * available.
+ * The (state, curr, stack) triple is used to save and restore parser
+ * state between async calls. State also helps encode extra
+ * information when streaming or unwrapping an array.
  *
- * The parser trait does not hold any state itself, but particular
- * implementations will usually hold state. Parser instances should
- * not be reused between parsing runs.
+ * The (data, len, allocated) triple is used to manage the underlying
+ * data the parser is keeping track of. As new data comes in, data may
+ * be expanded if not enough space is available.
  *
- * For now the parser requires input to be in UTF-8. This requirement
- * may eventually be relaxed.
+ * The offset parameter is used to drive the outer async parsing. It
+ * stores similar information to curr but is kept separate to avoid
+ * "corrupting" our snapshot.
+ *
+ * The done parameter is used internally to help figure out when the
+ * atEof() parser method should return true. This will be set when
+ * apply(None) is called.
+ *
+ * The streamMode parameter controls how the asynchronous parser will
+ * be handling multiple values. There are three states:
+ *
+ *    1: An array is being unwrapped. Normal JSON array rules apply
+ *       (Note that if the outer value observed is not an array, this
+ *       mode will toggle to the -1 mode).
+ *
+ *    0: A stream of individual JSON elements separated by whitespace
+ *       are being parsed. We can return each complete element as we
+ *       parse it.
+ *
+ *   -1: No streaming is occuring. Only a single JSON value is
+ *       allowed.
  */
-@SuppressWarnings(
-  Array(
-    "org.wartremover.warts.Var",
-    "org.wartremover.warts.FinalVal"))
-abstract class Parser[A](protected[this] final val plate: Plate[A]) {
+@SuppressWarnings(Array("org.wartremover.warts.Var"))
+final class Parser[A] protected[tectonic] (
+  _plate: Plate[A],
+  protected[tectonic] var state: Int,
+  protected[tectonic] var curr: Int,
+  protected[tectonic] var ring: Long,
+  protected[tectonic] var roffset: Int,
+  protected[tectonic] var fallback: BList,
+  protected[tectonic] var data: Array[Byte],
+  protected[tectonic] var len: Int,
+  protected[tectonic] var allocated: Int,
+  protected[tectonic] var offset: Int,
+  protected[tectonic] var done: Boolean,
+  protected[tectonic] var streamMode: Int
+) extends BaseParser[A](_plate) with ByteBasedParser[A] {
 
-  protected[this] final val utf8 = Charset.forName("UTF-8")
+  protected[this] var line = 0
+  protected[this] var pos = 0
+  protected[this] final def newline(i: Int): Unit = { line += 1; pos = i + 1 }
+  protected[this] final def column(i: Int) = i - pos
 
-  /**
-   * Read the byte/char at 'i' as a Char.
-   *
-   * Note that this should not be used on potential multi-byte
-   * sequences.
-   */
+  final def copy(): Parser[A] =
+    new Parser(_plate, state, curr, ring, roffset, fallback, data.clone, len, allocated, offset, done, streamMode)
+
+  @SuppressWarnings(
+    Array(
+      "org.wartremover.warts.NonUnitStatements",
+      "org.wartremover.warts.Overloading"))
+  final def absorb(buf: ByteBuffer): Either[ParseException, A] = {
+    done = false
+    val buflen = buf.limit() - buf.position()
+    val need = len + buflen
+    resizeIfNecessary(need)
+    buf.get(data, len, buflen)
+    len = need
+    churn()
+  }
+
   @SuppressWarnings(Array("org.wartremover.warts.Overloading"))
-  protected[this] def at(i: Int): Char
+  final def absorb(bytes: Array[Byte]): Either[ParseException, A] =
+    absorb(ByteBuffer.wrap(bytes))
 
-  /**
-   * Read the bytes/chars from 'i' until 'j' as a String.
-   */
   @SuppressWarnings(Array("org.wartremover.warts.Overloading"))
-  protected[this] def at(i: Int, j: Int): CharSequence
+  final def absorb(s: String): Either[ParseException, A] =
+    absorb(ByteBuffer.wrap(s.getBytes(utf8)))
 
-  /**
-   * Return true iff 'i' is at or beyond the end of the input (EOF).
-   */
-  protected[this] def atEof(i: Int): Boolean
+  final def finish(): Either[ParseException, A] = {
+    done = true
+    churn()
+  }
 
-  /**
-   * The reset() method is used to signal that we're working from the
-   * given position, and any previous data can be released. Some
-   * parsers (e.g.  StringParser) will ignore release, while others
-   * (e.g. PathParser) will need to use this information to release
-   * and allocate different areas.
-   */
-  protected[this] def reset(i: Int): Int
-
-  /**
-   * The checkpoint() method is used to allow some parsers to store
-   * their progress.
-   */
-  protected[this] def checkpoint(state: Int, i: Int, ring: Long, offset: Int, fallback: BList): Unit
-
-  /**
-   * Should be called when parsing is finished.
-   */
-  protected[this] def close(): Unit
-
-  /**
-   * Valid parser states.
-   */
-  @inline protected[this] final val ARRBEG = 6
-  @inline protected[this] final val OBJBEG = 7
-  @inline protected[this] final val DATA = 1
-  @inline protected[this] final val KEY = 2
-  @inline protected[this] final val SEP = 3
-  @inline protected[this] final val ARREND = 4
-  @inline protected[this] final val OBJEND = 5
-
-  protected[this] def newline(i: Int): Unit
-  protected[this] def line(): Int
-  protected[this] def column(i: Int): Int
-
-  @SuppressWarnings(Array("org.wartremover.warts.While"))
-  protected[this] final val HexChars: Array[Int] = {
-    val arr = new Array[Int](128)
-    var i = 0
-    while (i < 10) { arr(i + '0') = i; i += 1 }
-    i = 0
-    while (i < 16) { arr(i + 'a') = 10 + i; arr(i + 'A') = 10 + i; i += 1 }
-    arr
+  protected[this] final def resizeIfNecessary(need: Int): Unit = {
+    // if we don't have enough free space available we'll need to grow our
+    // data array. we never shrink the data array, assuming users will call
+    // feed with similarly-sized buffers.
+    if (need > allocated) {
+      val doubled = if (allocated < 0x40000000) allocated * 2 else Int.MaxValue
+      val newsize = max(need, doubled)
+      val newdata = new Array[Byte](newsize)
+      System.arraycopy(data, 0, newdata, 0, len)
+      data = newdata
+      allocated = newsize
+    }
   }
 
   /**
-   * Used to generate error messages with character info and offsets.
+   * Explanation of the new synthetic states. The parser machinery
+   * uses positive integers for states while parsing json values. We
+   * use these negative states to keep track of the async parser's
+   * status between json values.
+   *
+   * ASYNC_PRESTART: We haven't seen any non-whitespace yet. We
+   * could be parsing an array, or not. We are waiting for valid
+   * JSON.
+   *
+   * ASYNC_START: We've seen an array and have begun unwrapping
+   * it. We could see a ] if the array is empty, or valid JSON.
+   *
+   * ASYNC_END: We've parsed an array and seen the final ]. At this
+   * point we should only see whitespace or an EOF.
+   *
+   * ASYNC_POSTVAL: We just parsed a value from inside the array. We
+   * expect to see whitespace, a comma, or a ].
+   *
+   * ASYNC_PREVAL: We are in an array and we just saw a comma. We
+   * expect to see whitespace or a JSON value.
    */
+  @inline private[this] final def ASYNC_PRESTART = -5
+  @inline private[this] final def ASYNC_START = -4
+  @inline private[this] final def ASYNC_END = -3
+  @inline private[this] final def ASYNC_POSTVAL = -2
+  @inline private[this] final def ASYNC_PREVAL = -1
+
   @SuppressWarnings(
     Array(
-      "org.wartremover.warts.NonUnitStatements",
-      "org.wartremover.warts.Throw"))
-  protected[this] def die(i: Int, msg: String): Nothing = {
-    val y = line() + 1
-    val x = column(i) + 1
-    val s = "%s got %s (line %d, column %d)" format (msg, at(i), y, x)
-    throw ParseException(s, i, y, x)
-  }
+      "org.wartremover.warts.Equals",
+      "org.wartremover.warts.While"))
+  protected[tectonic] def churn(): Either[ParseException, A] = {
 
-  /**
-   * Used to generate messages for internal errors.
-   *
-   * This should only be used in situations where a possible bug in
-   * the parser was detected. For errors in user-provided JSON, use
-   * die().
-   */
-  protected[this] def error(msg: String) =
-    sys.error(msg)
+    // we rely on exceptions to tell us when we run out of data
+    try {
+      while (true) {
+        if (state < 0) {
+          (at(offset): @switch) match {
+            case '\n' =>
+              newline(offset)
+              offset += 1
 
-  /**
-   * Parse the given number, and add it to the given context.
-   *
-   * We don't actually instantiate a number here, but rather pass the
-   * string of for future use. Facades can choose to be lazy and just
-   * store the string. This ends up being way faster and has the nice
-   * side-effect that we know exactly how the user represented the
-   * number.
-   */
-  @SuppressWarnings(
-    Array(
-      "org.wartremover.warts.While",
-      "org.wartremover.warts.NonUnitStatements",
-      "org.wartremover.warts.Equals"))
-  protected[this] final def parseNum(i: Int): Int = {
-    var j = i
-    var c = at(j)
-    var decIndex = -1
-    var expIndex = -1
+            case ' ' | '\t' | '\r' =>
+              offset += 1
 
-    if (c == '-') {
-      j += 1
-      c = at(j)
-    }
-    if (c == '0') {
-      j += 1
-      c = at(j)
-    } else if ('1' <= c && c <= '9') {
-      while ('0' <= c && c <= '9') { j += 1; c = at(j) }
-    } else {
-      die(i, "expected digit")
-    }
+            case '[' =>
+              if (state == ASYNC_PRESTART) {
+                offset += 1
+                state = ASYNC_START
+              } else if (state == ASYNC_END) {
+                die(offset, "expected eof")
+              } else if (state == ASYNC_POSTVAL) {
+                die(offset, "expected , or ]")
+              } else {
+                state = 0
+              }
 
-    if (c == '.') {
-      decIndex = j - i
-      j += 1
-      c = at(j)
-      if ('0' <= c && c <= '9') {
-        while ('0' <= c && c <= '9') { j += 1; c = at(j) }
-      } else {
-        die(i, "expected digit")
-      }
-    }
+            case ',' =>
+              if (state == ASYNC_POSTVAL) {
+                offset += 1
+                state = ASYNC_PREVAL
+              } else if (state == ASYNC_END) {
+                die(offset, "expected eof")
+              } else {
+                die(offset, "expected json value")
+              }
 
-    if (c == 'e' || c == 'E') {
-      expIndex = j - i
-      j += 1
-      c = at(j)
-      if (c == '+' || c == '-') {
-        j += 1
-        c = at(j)
-      }
-      if ('0' <= c && c <= '9') {
-        while ('0' <= c && c <= '9') { j += 1; c = at(j) }
-      } else {
-        die(i, "expected digit")
-      }
-    }
+            case ']' =>
+              if (state == ASYNC_POSTVAL || state == ASYNC_START) {
+                if (streamMode > 0) {
+                  offset += 1
+                  state = ASYNC_END
+                } else {
+                  die(offset, "expected json value or eof")
+                }
+              } else if (state == ASYNC_END) {
+                die(offset, "expected eof")
+              } else {
+                die(offset, "expected json value")
+              }
 
-    plate.num(at(i, j), decIndex, expIndex)
-    j
-  }
-
-  /**
-   * Parse the given number, and add it to the given context.
-   *
-   * This method is a bit slower than parseNum() because it has to be
-   * sure it doesn't run off the end of the input.
-   *
-   * Normally (when operating in rparse in the context of an outer
-   * array or object) we don't need to worry about this and can just
-   * grab characters, because if we run out of characters that would
-   * indicate bad input. This is for cases where the number could
-   * possibly be followed by a valid EOF.
-   *
-   * This method has all the same caveats as the previous method.
-   */
-  @SuppressWarnings(
-    Array(
-      "org.wartremover.warts.Var",
-      "org.wartremover.warts.While",
-      "org.wartremover.warts.Return",
-      "org.wartremover.warts.NonUnitStatements",
-      "org.wartremover.warts.Equals"))
-  protected[this] final def parseNumSlow(i: Int): Int = {
-    var j = i
-    var c = at(j)
-    var decIndex = -1
-    var expIndex = -1
-
-    if (c == '-') {
-      // any valid input will require at least one digit after -
-      j += 1
-      c = at(j)
-    }
-    if (c == '0') {
-      j += 1
-      if (atEof(j)) {
-        plate.num(at(i, j), decIndex, expIndex)
-        return j
-      }
-      c = at(j)
-    } else if ('1' <= c && c <= '9') {
-      while ('0' <= c && c <= '9') {
-        j += 1
-        if (atEof(j)) {
-          plate.num(at(i, j), decIndex, expIndex)
-          return j
-        }
-        c = at(j)
-      }
-    } else {
-      die(i, "expected digit")
-    }
-
-    if (c == '.') {
-      // any valid input will require at least one digit after .
-      decIndex = j - i
-      j += 1
-      c = at(j)
-      if ('0' <= c && c <= '9') {
-        while ('0' <= c && c <= '9') {
-          j += 1
-          if (atEof(j)) {
-            plate.num(at(i, j), decIndex, expIndex)
-            return j
+            case c =>
+              if (state == ASYNC_END) {
+                die(offset, "expected eof")
+              } else if (state == ASYNC_POSTVAL) {
+                die(offset, "expected ] or ,")
+              } else {
+                if (state == ASYNC_PRESTART && streamMode > 0) streamMode = -1
+                state = 0
+              }
           }
-          c = at(j)
-        }
-      } else {
-        die(i, "expected digit")
-      }
-    }
 
-    if (c == 'e' || c == 'E') {
-      // any valid input will require at least one digit after e, e+, etc
-      expIndex = j - i
-      j += 1
-      c = at(j)
-      if (c == '+' || c == '-') {
-        j += 1
-        c = at(j)
-      }
-      if ('0' <= c && c <= '9') {
-        while ('0' <= c && c <= '9') {
-          j += 1
-          if (atEof(j)) {
-            plate.num(at(i, j), decIndex, expIndex)
-            return j
+        } else {
+          // jump straight back into rparse
+          offset = reset(offset)
+          val j = if (state <= 0) {
+            parse(offset)
+          } else {
+            rparse(state, curr, ring, roffset, fallback)
           }
-          c = at(j)
+          if (streamMode > 0) {
+            state = ASYNC_POSTVAL
+          } else if (streamMode == 0) {
+            state = ASYNC_PREVAL
+          } else {
+            state = ASYNC_END
+          }
+          curr = j
+          offset = j
         }
-      } else {
-        die(i, "expected digit")
       }
-    }
-
-    plate.num(at(i, j), decIndex, expIndex)
-    j
-  }
-
-  /**
-   * Generate a Char from the hex digits of "\u1234" (i.e. "1234").
-   *
-   * NOTE: This is only capable of generating characters from the basic plane.
-   * This is why it can only return Char instead of Int.
-   */
-  @SuppressWarnings(Array("org.wartremover.warts.While"))
-  protected[this] final def descape(s: CharSequence): Char = {
-    val hc = HexChars
-    var i = 0
-    var x = 0
-    while (i < 4) {
-      x = (x << 4) | hc(s.charAt(i).toInt)
-      i += 1
-    }
-    x.toChar
-  }
-
-  /**
-   * Parse the JSON string starting at 'i' and save it into the plate.
-   * If key is true, save the string with 'nestMap', otherwise use 'str'.
-   */
-  protected[this] def parseString(i: Int, key: Boolean): Int
-
-  /**
-   * Parse the JSON constant "true".
-   *
-   * Note that this method assumes that the first character has already been checked.
-   */
-  @SuppressWarnings(Array("org.wartremover.warts.Equals"))
-  protected[this] final def parseTrue(i: Int): Unit =
-    if (at(i + 1) == 'r' && at(i + 2) == 'u' && at(i + 3) == 'e') {
-      val _ = plate.tru()
-      ()
-    } else {
-      die(i, "expected true")
-    }
-
-  /**
-   * Parse the JSON constant "false".
-   *
-   * Note that this method assumes that the first character has already been checked.
-   */
-  @SuppressWarnings(Array("org.wartremover.warts.Equals"))
-  protected[this] final def parseFalse(i: Int): Unit =
-    if (at(i + 1) == 'a' && at(i + 2) == 'l' && at(i + 3) == 's' && at(i + 4) == 'e') {
-      val _ = plate.fls()
-      ()
-    } else {
-      die(i, "expected false")
-    }
-
-  /**
-   * Parse the JSON constant "null".
-   *
-   * Note that this method assumes that the first character has already been checked.
-   */
-  @SuppressWarnings(Array("org.wartremover.warts.Equals"))
-  protected[this] final def parseNull(i: Int): Unit =
-    if (at(i + 1) == 'u' && at(i + 2) == 'l' && at(i + 3) == 'l') {
-      val _ = plate.nul()
-      ()
-    } else {
-      die(i, "expected null")
-    }
-
-  /**
-   * Parse and return the next JSON value and the position beyond it.
-   */
-  @SuppressWarnings(
-    Array(
-      "org.wartremover.warts.Throw",
-      "org.wartremover.warts.Recursion",
-      "org.wartremover.warts.Null"))
-  protected[this] final def parse(i: Int): Int = try {
-    (at(i): @switch) match {
-      // ignore whitespace
-      case ' ' => parse(i + 1)
-      case '\t' => parse(i + 1)
-      case '\r' => parse(i + 1)
-      case '\n' => newline(i); parse(i + 1)
-
-      // if we have a recursive top-level structure, we'll delegate the parsing
-      // duties to our good friend rparse().
-      case '[' => rparse(ARRBEG, i + 1, 0L, 0, null)
-      case '{' => rparse(OBJBEG, i + 1, 1L, 0, null)
-
-      // we have a single top-level number
-      case '-' | '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' =>
-        val j = parseNumSlow(i)
-        plate.finishRow()
-        j
-
-      // we have a single top-level string
-      case '"' =>
-        val j = parseString(i, false)
-        plate.finishRow()
-        j
-
-      // we have a single top-level constant
-      case 't' =>
-        parseTrue(i)
-        plate.finishRow()
-        i + 4
-
-      case 'f' =>
-        parseFalse(i)
-        plate.finishRow()
-        i + 5
-
-      case 'n' =>
-        parseNull(i)
-        plate.finishRow()
-        i + 4
-
-      // invalid
-      case _ => die(i, "expected json value")
-    }
-  } catch {
-    case _: IndexOutOfBoundsException =>
-      throw IncompleteParseException("exhausted input")
-  }
-
-  /**
-   * Tail-recursive parsing method to do the bulk of JSON parsing.
-   *
-   * This single method manages parser states, data, etc. Except for
-   * parsing non-recursive values (like strings, numbers, and
-   * constants) all important work happens in this loop (or in methods
-   * it calls, like reset()).
-   *
-   * Currently the code is optimized to make use of switch
-   * statements. Future work should consider whether this is better or
-   * worse than manually constructed if/else statements or something
-   * else. Also, it may be possible to reorder some cases for speed
-   * improvements.
-   */
-  @tailrec
-  @SuppressWarnings(
-    Array(
-      "org.wartremover.warts.NonUnitStatements",
-      "org.wartremover.warts.Equals"))
-  protected[this] final def rparse(state: Int, j: Int, ring: Long, offset: Int, fallback: BList): Int = {
-    val i = reset(j)
-    checkpoint(state, i, ring, offset, fallback)
-
-    val c = at(i)
-
-    if (c == '\n') {
-      newline(i)
-      rparse(state, i + 1, ring, offset, fallback)
-    } else if (c == ' ' || c == '\t' || c == '\r') {
-      rparse(state, i + 1, ring, offset, fallback)
-    } else if (state == DATA) {
-      // we are inside an object or array expecting to see data
-      if (c == '[') {
-        var offset2 = offset
-        var ring2 = ring
-        var fallback2 = fallback
-
-        if (checkPushEnclosure(ring, offset, fallback)) {
-          offset2 = offset + 1
-          ring2 = pushEnclosureRing(ring, offset, false)
+      Right(plate.finishBatch(false))
+    } catch {
+      case e: AsyncException =>
+        if (done) {
+          // if we are done, make sure we ended at a good stopping point
+          if (state == ASYNC_PREVAL || state == ASYNC_END) Right(plate.finishBatch(true))
+          else Left(ParseException("exhausted input", -1, -1, -1))
         } else {
-          fallback2 = pushEnclosureFallback(fallback, false)
+          // we ran out of data, so return what we have so far
+          Right(plate.finishBatch(false))
         }
 
-        rparse(ARRBEG, i + 1, ring2, offset2, fallback2)
-      } else if (c == '{') {
-        var offset2 = offset
-        var ring2 = ring
-        var fallback2 = fallback
+      case e: ParseException =>
+        // we hit a parser error, so return that error and results so far
+        Left(e)
+    }
+  }
 
-        if (checkPushEnclosure(ring, offset, fallback)) {
-          offset2 = offset + 1
-          ring2 = pushEnclosureRing(ring, offset, true)
-        } else {
-          fallback2 = pushEnclosureFallback(fallback, true)
-        }
-
-        rparse(OBJBEG, i + 1, ring2, offset2, fallback2)
-      } else {
-        if ((c >= '0' && c <= '9') || c == '-') {
-          val j = parseNum(i)
-          rparse(if (enclosure(ring, offset, fallback)) OBJEND else ARREND, j, ring, offset, fallback)
-        } else if (c == '"') {
-          val j = parseString(i, false)
-          rparse(if (enclosure(ring, offset, fallback)) OBJEND else ARREND, j, ring, offset, fallback)
-        } else if (c == 't') {
-          parseTrue(i)
-          rparse(if (enclosure(ring, offset, fallback)) OBJEND else ARREND, i + 4, ring, offset, fallback)
-        } else if (c == 'f') {
-          parseFalse(i)
-          rparse(if (enclosure(ring, offset, fallback)) OBJEND else ARREND, i + 5, ring, offset, fallback)
-        } else if (c == 'n') {
-          parseNull(i)
-          rparse(if (enclosure(ring, offset, fallback)) OBJEND else ARREND, i + 4, ring, offset, fallback)
-        } else {
-          die(i, "expected json value")
-        }
-      }
-    } else if (
-      (c == ']' && (state == ARREND || state == ARRBEG)) ||
-      (c == '}' && (state == OBJEND || state == OBJBEG))
-    ) {
-      // we are inside an array or object and have seen a key or a closing
-      // brace, respectively.
-
-      var offset2 = offset
-      var fallback2 = fallback
-
-      if (checkPopEnclosure(ring, offset, fallback))
-        offset2 = offset - 1
-      else
-        fallback2 = popEnclosureFallback(fallback)
-
-      (state: @switch) match {
-        case ARRBEG => plate.arr()
-        case OBJBEG => plate.map()
-        case ARREND | OBJEND => plate.unnest()
-      }
-
-      if (offset2 < 0) {
-        plate.finishRow()
-        i + 1
-      } else {
-        rparse(if (enclosure(ring, offset2, fallback2)) OBJEND else ARREND, i + 1, ring, offset2, fallback2)
-      }
-    } else if (state == KEY) {
-      // we are in an object expecting to see a key.
-      if (c == '"') {
-        rparse(SEP, parseString(i, true), ring, offset, fallback)
-      } else {
-        die(i, "expected \"")
-      }
-    } else if (state == SEP) {
-      // we are in an object just after a key, expecting to see a colon.
-      if (c == ':') {
-        rparse(DATA, i + 1, ring, offset, fallback)
-      } else {
-        die(i, "expected :")
-      }
-    } else if (state == ARREND) {
-      // we are in an array, expecting to see a comma (before more data).
-      if (c == ',') {
-        plate.unnest()
-        plate.nestArr()
-        rparse(DATA, i + 1, ring, offset, fallback)
-      } else {
-        die(i, "expected ] or ,")
-      }
-    } else if (state == OBJEND) {
-      // we are in an object, expecting to see a comma (before more data).
-      if (c == ',') {
-        plate.unnest()
-        rparse(KEY, i + 1, ring, offset, fallback)
-      } else {
-        die(i, "expected } or ,")
-      }
-    } else if (state == ARRBEG) {
-      // we are starting an array, expecting to see data or a closing bracket.
-      plate.nestArr()
-      rparse(DATA, i, ring, offset, fallback)
+  // every 1M we shift our array back to the beginning.
+  protected[this] final def reset(i: Int): Int = {
+    if (offset >= 1048576) {
+      val diff = offset
+      curr -= diff
+      len -= diff
+      offset = 0
+      pos -= diff
+      System.arraycopy(data, diff, data, 0, len)
+      i - diff
     } else {
-      // we are starting an object, expecting to see a key or a closing brace.
-      rparse(KEY, i, ring, offset, fallback)
+      i
     }
   }
 
   /**
-   * A value of true indicates an object, and false indicates an array. Note that
-   * a non-existent enclosure is indicated by offset < 0
+   * We use this to keep track of the last recoverable place we've
+   * seen. If we hit an AsyncException, we can later resume from this
+   * point.
+   *
+   * This method is called during every loop of rparse, and the
+   * arguments are the exact arguments we can pass to rparse to
+   * continue where we left off.
    */
-  @inline
-  @SuppressWarnings(Array("org.wartremover.warts.Equals"))
-  private[this] def enclosure(ring: Long, offset: Int, fallback: BList): Boolean = {
-    if (fallback == null)
-      (ring & (1L << offset)) != 0
-    else
-      fallback.head
+  protected[this] final def checkpoint(state: Int, i: Int, ring: Long, offset: Int, fallback: BList): Unit = {
+    this.state = state
+    this.curr = i
+    this.ring = ring
+    this.roffset = offset
+    this.fallback = fallback
   }
 
-  @inline
-  @SuppressWarnings(Array("org.wartremover.warts.Equals"))
-  private[this] def checkPushEnclosure(ring: Long, offset: Int, fallback: BList): Boolean =
-    fallback == null
+  /**
+   * This is a specialized accessor for the case where our underlying data are
+   * bytes not chars.
+   */
+  @SuppressWarnings(Array("org.wartremover.warts.Throw"))
+  protected[this] final def byte(i: Int): Byte =
+    if (i >= len) throw new AsyncException else data(i)
 
-  @inline
-  private[this] def pushEnclosureRing(ring: Long, offset: Int, enc: Boolean): Long = {
-    if (enc)
-      ring | (1L << (offset + 1))
-    else
-      ring & ~(1L << (offset + 1))
+  // we need to signal if we got out-of-bounds
+  @SuppressWarnings(Array("org.wartremover.warts.Throw"))
+  protected[this] final def at(i: Int): Char =
+    if (i >= len) throw new AsyncException else data(i).toChar
+
+  /**
+   * Access a byte range as a string.
+   *
+   * Since the underlying data are UTF-8 encoded, i and k must occur on unicode
+   * boundaries. Also, the resulting String is not guaranteed to have length
+   * (k - i).
+   */
+  @SuppressWarnings(Array("org.wartremover.warts.Throw"))
+  protected[this] final def at(i: Int, k: Int): CharSequence = {
+    if (k > len) throw new AsyncException
+    val size = k - i
+    val arr = new Array[Byte](size)
+    System.arraycopy(data, i, arr, 0, size)
+    new String(arr, utf8)
   }
 
-  @inline
-  private[this] def pushEnclosureFallback(fallback: BList, enc: Boolean): BList =
-    enc :: fallback
+  // the basic idea is that we don't signal EOF until done is true, which means
+  // the client explicitly send us an EOF.
+  protected[this] final def atEof(i: Int): Boolean =
+    if (done) i >= len else false
 
-  @inline
-  @SuppressWarnings(Array("org.wartremover.warts.Equals"))
-  private[this] def checkPopEnclosure(ring: Long, offset: Int, fallback: BList): Boolean =
-    fallback == null
-
-  @inline
-  @SuppressWarnings(
-    Array(
-      "org.wartremover.warts.IsInstanceOf",
-      "org.wartremover.warts.AsInstanceOf",
-      "org.wartremover.warts.Null"))
-  private[this] def popEnclosureFallback(fallback: BList): BList = {
-    if (fallback.isInstanceOf[BList.Last])
-      null
-    else
-      fallback.asInstanceOf[BList.Cons].tail
-  }
+  // we don't have to do anything special on close.
+  protected[this] final def close(): Unit = ()
 }
+
+/**
+ * This class is used internally by Parser to signal that we've
+ * reached the end of the particular input we were given.
+ */
+private[tectonic] class AsyncException extends Exception with control.NoStackTrace
+
+/**
+ * This is a more prosaic exception which indicates that we've hit a
+ * parsing error.
+ */
+private[tectonic] class FailureException extends Exception
